Create HasRelation concern
