+++ packages/tables/src/Filters/MultiSelectFilter.php
@@ -5,15 +5,12 @@ namespace Filament\Tables\Filters;
 use Closure;
 use Filament\Forms\Components\MultiSelect;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\MorphToMany;
-use Illuminate\Database\Eloquent\Relations\Relation;
-use Illuminate\Support\Str;
 
 class MultiSelectFilter extends Filter
 {
     use Concerns\HasOptions;
     use Concerns\HasPlaceholder;
+    use Concerns\HasRelation;
 
     protected string | Closure | null $column = null;
 
@@ -43,18 +40,10 @@ class MultiSelectFilter extends Filter
         if ($this->queriesRelationships()) {
             $relationship = $this->getRelationship();
 
-            if ($relationship instanceof MorphToMany) {
-                /** @var MorphToMany $relationship */
-                $keyColumn = $relationship->getParentKeyName();
-            } else {
-                /** @var BelongsTo $relationship */
-                $keyColumn = $relationship->getOwnerKeyName();
-            }
-
             return $query->whereHas(
                 $this->getRelationshipName(),
                 fn (Builder $query) => $query->whereIn(
-                    $keyColumn,
+                    $this->getRelationKey($relationship),
                     $data['values'],
                 ),
             );
@@ -73,13 +62,6 @@ class MultiSelectFilter extends Filter
         return $this;
     }
 
-    public function relationship(string $relationshipName, string $displayColumnName = null): static
-    {
-        $this->column("{$relationshipName}.{$displayColumnName}");
-
-        return $this;
-    }
-
     public function static(bool | Closure $condition = true): static
     {
         $this->isStatic = $condition;
@@ -92,27 +74,6 @@ class MultiSelectFilter extends Filter
         return $this->evaluate($this->column) ?? $this->getName();
     }
 
-    protected function getRelationshipOptions(): array
-    {
-        $relationship = $this->getRelationship();
-
-        if ($relationship instanceof MorphToMany) {
-            /** @var MorphToMany $relationship */
-            $keyColumn = $relationship->getParentKeyName();
-        } else {
-            /** @var BelongsTo $relationship */
-            $keyColumn = $relationship->getOwnerKeyName();
-        }
-
-        $displayColumnName = $this->getRelationshipDisplayColumnName();
-
-        $relationshipQuery = $relationship->getRelated()->query()->orderBy($displayColumnName);
-
-        return $relationshipQuery
-            ->pluck($displayColumnName, $keyColumn)
-            ->toArray();
-    }
-
     public function getFormSchema(): array
     {
         return $this->formSchema ?? [
@@ -122,27 +83,5 @@ class MultiSelectFilter extends Filter
                 ->placeholder($this->getPlaceholder())
                 ->default($this->getDefaultState()),
         ];
-    }
-
-    public function queriesRelationships(): bool
-    {
-        return Str::of($this->getColumn())->contains('.');
-    }
-
-    protected function getRelationship(): Relation
-    {
-        $model = app($this->getTable()->getModel());
-
-        return $model->{$this->getRelationshipName()}();
-    }
-
-    protected function getRelationshipName(): string
-    {
-        return (string) Str::of($this->getColumn())->beforeLast('.');
-    }
-
-    protected function getRelationshipDisplayColumnName(): string
-    {
-        return (string) Str::of($this->getColumn())->afterLast('.');
     }
 }
