+++ packages/tables/src/Filters/SelectFilter.php
@@ -5,15 +5,12 @@ namespace Filament\Tables\Filters;
 use Closure;
 use Filament\Forms\Components\Select;
 use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\MorphToMany;
-use Illuminate\Database\Eloquent\Relations\Relation;
-use Illuminate\Support\Str;
 
 class SelectFilter extends Filter
 {
     use Concerns\HasOptions;
     use Concerns\HasPlaceholder;
+    use Concerns\HasRelation;
 
     protected string | Closure | null $column = null;
 
@@ -45,17 +42,9 @@ class SelectFilter extends Filter
         if ($this->queriesRelationships()) {
             $relationship = $this->getRelationship();
 
-            if ($relationship instanceof MorphToMany) {
-                /** @var MorphToMany $relationship */
-                $keyColumn = $relationship->getParentKeyName();
-            } else {
-                /** @var BelongsTo $relationship */
-                $keyColumn = $relationship->getOwnerKeyName();
-            }
-
             return $query->whereRelation(
                 $this->getRelationshipName(),
-                $keyColumn,
+                $this->getRelationKey($relationship),
                 $data['value'],
             );
         }
@@ -66,13 +55,6 @@ class SelectFilter extends Filter
     public function column(string | Closure | null $name): static
     {
         $this->column = $name;
-
-        return $this;
-    }
-
-    public function relationship(string $relationshipName, string $displayColumnName = null): static
-    {
-        $this->column("{$relationshipName}.{$displayColumnName}");
 
         return $this;
     }
@@ -96,27 +78,6 @@ class SelectFilter extends Filter
         return $this->evaluate($this->column) ?? $this->getName();
     }
 
-    protected function getRelationshipOptions(): array
-    {
-        $relationship = $this->getRelationship();
-
-        if ($relationship instanceof MorphToMany) {
-            /** @var MorphToMany $relationship */
-            $keyColumn = $relationship->getParentKeyName();
-        } else {
-            /** @var BelongsTo $relationship */
-            $keyColumn = $relationship->getOwnerKeyName();
-        }
-
-        $displayColumnName = $this->getRelationshipDisplayColumnName();
-
-        $relationshipQuery = $relationship->getRelated()->query()->orderBy($displayColumnName);
-
-        return $relationshipQuery
-            ->pluck($displayColumnName,  $keyColumn)
-            ->toArray();
-    }
-
     public function getFormSchema(): array
     {
         return $this->formSchema ?? [
@@ -132,27 +93,5 @@ class SelectFilter extends Filter
     public function isSearchable(): bool
     {
         return (bool) $this->evaluate($this->isSearchable);
-    }
-
-    public function queriesRelationships(): bool
-    {
-        return Str::of($this->getColumn())->contains('.');
-    }
-
-    protected function getRelationship(): Relation
-    {
-        $model = app($this->getTable()->getModel());
-
-        return $model->{$this->getRelationshipName()}();
-    }
-
-    protected function getRelationshipName(): string
-    {
-        return (string) Str::of($this->getColumn())->beforeLast('.');
-    }
-
-    protected function getRelationshipDisplayColumnName(): string
-    {
-        return (string) Str::of($this->getColumn())->afterLast('.');
     }
 }
